Added file cache support to avoid the initial build overhead.
When retrieving cache entries, the config file and its list of file dependencies will be checked for any changes since the compilation. If all the output is still up to date, the cached output is served.

When watching, the cache is used only until the watcher has completed the first build.