Added checks to prevent multiple mutations on a config file.
A process will now produce errors if a request comes in for a build of a config file which has already been imported by another build. The `buildHash` prop is used as a unique identifier to detect multiple requests.

Workers can now indicate if they can safely handle a build, which enables builds to be passed to the next available _and_ capable worker